+++ source/ChanSort/Printing/ReportOptionsDialog.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Drawing;
 using ChanSort.Api;
+using ChanSort.Ui.Properties;
 using DevExpress.LookAndFeel;
 using DevExpress.XtraEditors;
 using DevExpress.XtraReports.UI;
@@ -17,18 +18,18 @@ namespace ChanSort.Ui.Printing
       InitializeComponent();
       this.channelList = channelList;
       this.subListIndex = subListIndex;
-      if (Properties.Settings.Default.PrintSortByName)
+      if (Config.Default.PrintSortByName)
         this.rbSortByName.Checked = true;
-      this.fontEdit1.EditValue = Properties.Settings.Default.PrintFontName;
-      this.spinFontSize.Value = Properties.Settings.Default.PrintFontSize;
+      this.fontEdit1.EditValue = Config.Default.PrintFontName;
+      this.spinFontSize.Value = Config.Default.PrintFontSize;
     }
 
     private void btnPreview_Click(object sender, EventArgs e)
     {
-      Properties.Settings.Default.PrintFontName = (string)this.fontEdit1.EditValue;
-      Properties.Settings.Default.PrintFontSize = this.spinFontSize.Value;
-      Properties.Settings.Default.PrintSortByName = this.rbSortByNumber.Checked;
-      Properties.Settings.Default.Save();
+      Config.Default.PrintFontName = (string)this.fontEdit1.EditValue;
+      Config.Default.PrintFontSize = this.spinFontSize.Value;
+      Config.Default.PrintSortByName = this.rbSortByNumber.Checked;
+      Config.Default.Save();
 
       using (var font = new Font(this.fontEdit1.Text, (float)this.spinFontSize.Value))
       using (var report = new ChannelListReport(this.channelList, this.subListIndex, this.rbSortByName.Checked, font))